requirements.txt
requests==2.31.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
python-dotenv==1.0.1
matplotlib==3.8.2
seaborn==0.13.2
scipy==1.12.0
jupyter==1.0.0
ipykernel==6.29.2
rich>=13.7.0
schedule>=1.2.0
python-telegram-bot>=22.0
fastapi
uvicorn
jinja2
stripe
resend
python-jose
passlib
bcrypt
sqlalchemy
python-multipart
apscheduler>=3.10.0
anthropic>=0.26.0
psycopg2-binary>=2.9.0
